# contracts/openapi/pramaan-backend.yaml
openapi: 3.1.0
info:
  title: Pramaan Next — Backend (Phase 1)
  version: 0.1.0
  description: >
    Evidence ingestion, versioning, integrity, deterministic repository queries,
    application metadata, and the collection scheduler. Phase 1 uses deterministic
    mock enterprise integrations. No LLM. Source of truth for backend ⇄ frontend
    and backend ⇄ agents interfaces — implement against this file, do not hand-roll.
servers:
  - url: http://localhost:8080
    description: Local dev
...
paths:
...
  /api/v1/scheduler/runs:
    post:
      tags: [scheduler]
      summary: Start a collection run (async). Empty body = all apps + default sources.
      requestBody:
        required: false
        content: { application/json: { schema: { $ref: "#/components/schemas/RunRequest" } } }
      responses:
        "202": { description: Accepted, content: { application/json: { schema: { $ref: "#/components/schemas/RunView" } } } }
    get:
      tags: [scheduler]
      summary: List runs (newest first)
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/RunView" } }
  /api/v1/scheduler/runs/{runId}:
    parameters: [ { name: runId, in: path, required: true, schema: { type: string, format: uuid } } ]
    get:
      tags: [scheduler]
      summary: Run status / progress
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/RunView" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/scheduler/runs/{runId}/retry:
    parameters: [ { name: runId, in: path, required: true, schema: { type: string, format: uuid } } ]
    post:
      tags: [scheduler]
      summary: Re-run a finished run with the same parameters
      responses:
        "202": { description: Accepted, content: { application/json: { schema: { $ref: "#/components/schemas/RunView" } } } }
        "409": { $ref: "#/components/responses/Error" }

  # ---- Use Case 1 — evidence detail context -----------------------------
...
components:
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
  schemas:
    ErrorBody:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
    Page:
      type: object
      properties:
        items: { type: array, items: {} }
        page: { type: integer }
        size: { type: integer }
        totalItems: { type: integer, format: int64 }
        totalPages: { type: integer }
...
    RunRequest:
      type: object
      properties:
        applications: { type: array, items: { type: string } }
        frameworks: { type: array, items: { type: string } }
        sources: { type: array, items: { type: string } }
        requestedBy: { type: string }
    RunView:
      type: object
      properties:
        runId: { type: string, format: uuid }
        trigger: { type: string, enum: [MANUAL, SCHEDULED, BULK] }
        status: { type: string, enum: [PENDING, RUNNING, COMPLETED, FAILED] }
        requestedBy: { type: string }
        applications: { type: array, items: { type: string } }
        frameworks: { type: array, items: { type: string } }
        sources: { type: array, items: { type: string } }
        received: { type: integer }
        ingested: { type: integer }
        duplicates: { type: integer }
        failed: { type: integer }
        message: { type: string }
        perSource: { type: object, additionalProperties: { type: string } }
        createdAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time, nullable: true }
        finishedAt: { type: string, format: date-time, nullable: true }

    # ---- Use Case 1 — evidence context ----
...